Warrior model

+++ src/main/java/ar/edu/itba/sia/model/Chromosome.java
@@ -2,7 +2,7 @@ package ar.edu.itba.sia.model;
 
 public interface Chromosome {
 
-    int fitness();
+    double fitness();
 
     double relativeFitness();
