+++ src/main/java/ar/edu/itba/sia/selection/Rank.java
@@ -12,7 +12,7 @@ public class Rank implements Selection {
     @Override
     public List<Chromosome> select(int amount, List<Chromosome> population) {
         if (amount < 0 || amount > population.size()) throw new IllegalArgumentException();
-        Collections.sort(population, (a, b) -> a.fitness() - b.fitness()); // Order by fitness, worst to best
+        Collections.sort(population, (a, b) -> (int) (a.fitness() - b.fitness())); // Order by fitness, worst to best
 
         i = 1;
         float populationSize = population.size();
